+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ### Added
 
-- Get Notification status for Pull Request from Details
+- Get Notification status for Pull Request and Issue from Details
 
 ## [0.4.3] - 2025-01-18